Start refactoring the failover code.
Use a static carp-state-change hook file, that is python.  Stop
generating a shell script from a shell script all the time.
This will get generated by ix-devd called by changing anything in the GUI
or by hactl that has anything to do with failover.

This is not the python version of carp-state-change that reads a json
file generated by ix-devd.  This is just the thing that generates the
json file that will get read by the rewritten carp-state-change file.

Ticket: #7281