Flake fixes
I made the previous mistake in fixing flake errors the
CI didn't seem to care too much about, so my careful
reduction in flake errors was wasted.

In this pass I'm fixing flake errors in a file this
pull request doesn't touch, but the FreeNAS CI is
specifically complaining about, so hopefully this will
do the trick.

Also, I've gone a bit overboard and actually reduced the
number of flake errors in the project now.  Hoefully that
doesn't violate some law of conservation of flake errors
and create a wormhole, naked singularity, or Cthulu Summoning
that summons ruin on everything we hold dear in this world.