Configure locking/setting passphrase correctly
This commit introduces changes which make sure that we don't lock a pool which is not being used by system dataset service or allow the end user to set a passphrase for a pool which is being used by system dataset service. For any users who already do this, we don't allow them to change their passphrase for the selected pool but only to remove it as long as that pool is being used by system dataset service.
Ticket: #71971