Detect EFI system partition when upgrading loader
For in-place upgrades, the boot type of the existing installation is
detected to ensure the correct type of boot code is installed.  When
GRUB was the boot loader, installations labelled the ESP 'efibsd',
however the new bsdloader installations do not use a label.  Both
installation styles use the 'efi' partition type, so check for that
instead of the 'efibsd' label to ensure we do not convert the UEFI
system partition to a broken legacy BIOS boot partition.

Ticket: NAS-101442