Properly restart UPS
When UPS service is restarted, upsmon daemon in some cases has not stopped it's multiple processes entirely and is in the process of doing so when we initiate a restart. Restart fails in this case as old process is still cleaning itself. This commit makes sure we properly wait for upsmon to die and then initiate a restart.