Validate Client Config
This commit adds a method which can be called before we start OpenVPN client making sure that OpenVPN Client does not run with a misconfigured file resulting in further woes for the user. This will raise a CallError if the settings are not as they should be for OpenVPN client and we can skip generating the openvpn client file which will result in failure to start for the service.