Ability to create plugins with plugin service
This commit exposes ability to create plugins with plugin service. While doing this, we enforce the consumer of api to provide a jail name to be deterministic about the jail's name instead of allowing iocage to allocate a name dynamically which results in no solid way of identifying which jail was recently created.