Bug fix for an unreferenced variable (#3298)
Remove notifier.directoryservice method
This method should not be used now that the directory services are moved to middleware. The only place we used this method was where we were generating the afp.conf file. Move the relevant logic to the etc file and use the proper new API calls. Also add a relatively detailed comment about why we're putting this information in the afp.conf file in some situations.