NAS-102533 / 11.3 / Fix API for smb.sharesec (#3339)
* Fix API for smb.sharesec
Convert sharesec into a CRUD service. When _view_all is called (for instance through smb.sharesec.query), add a numerical index key to the parsed security descritor for each share. Since the ordering is fairly consistent, this can serve as a unique ID for the update and delete methods. Delete can take the numeric id or share name as an argument. Deletion in this case means replacing the existing Share ACL with S-1-1-0/FULL. The distinction between create and update is that the former takes a share name and the latter a numeric ID since it is only possible to apply an ACL to an existing share.