Migrate "unified" AD configuration
TrueNAS 11.1 had an undocumented kludge in place that would allow support
to use a single computer object in AD if the netbios names of both controllers
were identical. This is now the default configuration in 11.2, but as a
side-effect, the virtual hostname _must_ be set.

(cherry picked from commit 00e19ce77b5e3229e322fe90dd3f60bf8ff00ccc)