Validate directory service config during health checks
When LDAP or AD is enabled, certain assumptions are made about the
service having valid data in the service config. Since the presence
of bad data in the config can adversely affect system behavior
outside of the service in question (and there are some online guides
suggesting using the sqlite command to manually edit the database),
perform an additional configuration validation step the respective
<ds>.started methods. These methods are called in periodic health
checks as part of the directory service alerts, and will convert
the ValidationError into a proper alert. When this check fails,
disable the directory service in question to help recover from
undefined behavior.