Fix collectd starting with default config
Now etc.generate_all will clear it on system boot

Before services rewrite, disk.restart_services_after_sync
was doing the same for us.

It was doing so because call('service.started', 'collectd') was
erroneously returning true.

It was returning true because it was doing `service collectd-daemon quietstatus`
which was returning 0 because rc variable was still not configured at that moment.