Gracefully handle unexpected LDAP schemas (#4641)
Windows LDAP servers (AD environment) have slightly different return to query
for SambaDomain than we see on OpenLDAP server. Log a debug message with the
contents that we failed to parse and then move on. In the future, it may make
sense to add a validation error to redirect users to the AD form if a query to
rootDSE shows it's an AD DC.