Introduce WAITING truecommand state
This commit introduces a new state for TC service 'WAITING'. Motivation is that if TC has not connected to NAS or if the last time it did was more then 30 min ago, we change NAS TC service state to WAITING which shows that the NAS is waiting for TC to connect.