Ping TC IP address if handshake check is successful
This commit introduces changes where we ping TC ip address to determine if we have an active connection with TC. Why we do this is because it's possible TC container changed it's IP address and the keys are still valid, which means we would get a handshake from the wireguard connection but we are not connected to the TC container anymore. In this case we should get in an error state and poll the portal to see if we should be using updated address for the TC container.