installer: Fix uefi serial first boot
When we have booted with only a serial console, enable only comconsole
in loader.conf.local for the first boot, to avoid potentially doubling
the loader output when efi is a serial console (such as in bhyve).

Middleware correctly generates loader.conf.local with only comconsole
enabled on subsequent boots.