install.sh: Fix fresh installs to a BE
This was a case that had been overlooked while cleaning up and
restructuring the installer.  We were formatting for any install that
was not an inplace upgrade.

Factor out a check for if we're doing an upgrade vs fresh install to
make things a bit more readable.  Use it to rewrite some of the logic
for deciding how to proceed during installation.  Don't format the disks
when doing an inplace fresh install to a BE.

While here, fix a bug that was encountered in testing the patch where an
uninitialized variable was reused when starting over the installer,
causing disks to appear twice in the disk list.

Ticket: https://jira.ixsystems.com/browse/NAS-106777
(cherry picked from commit 576b53c280d66db10f67f42f6fec613e198a84a9)