Correctly handle ACME certs using SAN
When domain names are specified for SAN, we prefix them with 'DNS:' following with the cert standards, during acme authorisation, ACME server strips them and sends the authorisation for the domain name. We should handle this gracefully and strip out the prefix on our end as well confirming authorisation for the ACME order.

(cherry picked from commit 7fcafafc7015427ab3920c015c8dd43748563cfb)