Donot raise UPS alert during self test
There are some UPS models which run a self test at scheduled intervals where they go on battery and back during the test. We should not raise an alert in this case as this is scheduled and expected and gives a false alarm.

(cherry picked from commit d358641b7da585ae253a4fe56ae05153b764265d)