freenas-debug: replace pdbedit with smb.passdb_list
When the LDAP directory service is enabled with a samba schema, pdbedit
-L will perform synchronous LDAP queries to resolve every user in the
LDAP environment. This is somewhat undesireable. Replace this with a
call to middleware's smb.passdb_list, which will return an empty list if
the passdb backend is ldapsam.

(cherry picked from commit 3343773dc21f6dd8d74bac87bada8d77d9ccd276)