Try enabling capabilities at once for interfaces
There are interface drivers where some capability is dependant on some other capability so if try to enable X capability which is dependant on Y, it will not be enabled unless both of them are enabled together. After discussing with mav, we should first try to enable capabilities together and if that fails we can revert to trying to enable them one by one.