Do not use bhyve args for VNC device
This commit adds changes to not use bhyve args for VNC device as they are broken wrt VNC device with VNC server not respecting the specified port configuration resulting in VNC device not working. Changes have been made to instead use the official/proper way of specifying VNC device which earlier was not being used as it did not support resolution/password/wait attribute of vnc device which now it does except for wait attr.