NAS-115027 / 22.12 / Optimize reloads of user service (#8386)
* Optimize reloads of user service

Consolidate all the mako files associated with a user service
reload into a single python file. This minimizes the amount
of times that user.query and group.query must be called. On
test server with high user count, this changeset halves the
amount of time it takes to add or delete a new user.

* Enhance user API test

Improve existing user tests by validating the following:
- pw_dir reflects the homedir specified when creating a user
- pw_gecos reflects the specified user

Verify that /etc/master.passwd, /etc/group, /etc/sudoers
are properly written in a way that reflects account modifications.